Módulo 05 - Aula 24

+++ course-01/module-05/classroom-24/Program.cs
@@ -2,9 +2,10 @@
 {
     private static void Main(string[] args)
     {
-        int inteiro = int.Parse("100"); // Forma correta
+        decimal inteiro = decimal.Parse("100"); // Forma correta
         //int inteiro = int.Parse("ABC"); // Forma errada
 
         Console.WriteLine($"\nForma correta: {inteiro}\n");
     }
 }
+
